+++ middleware.ts
@@ -17,5 +17,5 @@ export default auth((req) => {
 })
 
 export const config = {
-  matcher: ["/((?!api/auth|_next/static|_next/image|favicon.ico|manifest.webmanifest|icon-192|icon-512).*)"],
+  matcher: ["/((?!api/auth|api/login|_next/static|_next/image|favicon.ico|manifest.webmanifest|icon-192|icon-512).*)"],
 }
